refactor: replace roundToLong().toInt() with roundToInt() and remove redundant toLong()

+++ banchus-common/src/main/kotlin/pe/nanamochi/banchus/database/entity/Channel.kt
@@ -47,7 +47,7 @@ sealed class ChannelName {
     fun getUpdateStream(): StreamName {
         return when (this) {
             is Spectator -> StreamName.Spectator(this.sessionId)
-            is Multiplayer -> StreamName.Multiplayer(this.matchId.toLong())
+            is Multiplayer -> StreamName.Multiplayer(this.matchId)
             is Chat -> {
                 when (this.name) {
                     "#plus",
